move NewSpendingLimitDetails under TxDescription umbrella
